desktop: Move parse_url and pick_file to util.rs
